Rename component to Alert and made it more generic
